requirements.txt
aiogram==3.22.0
mysql-connector-python==9.4.0
python-dotenv==1.1.1
aiohttp==3.12.15




